requirements.txt
asgiref==3.4.1
cffi==1.15.1
cryptography==38.0.3
Django==3.2.16
djangorestframework==3.14.0
numpy==1.19.5
Pillow==8.4.0
pycparser==2.21
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
